requirements.txt
altair==4.0.1
numpy==1.18.1
Pillow==7.0.0
regex==2017.4.5
requests==2.23.0
streamlit==0.56.0
